remove unused 'use client' directive and clean up Xps component data structure
